requirements.txt
streamlit
python-docx
openai
tiktoken 